+++ internal/db/migration/000001_init_schema.up.sql
