+++ internal/db/query/session.sql
