+++ internal/utils/ethereum_wallet_test.go
